@DingDongSoLong4 2022-09-28
Oops, forgot about NULLs being a pain in SQL. Should be fixed now.